refactor: reorganize kubectl and internal API handlers
Move handler modules from server root to organized utils subdirectories:
- kubectl_agent_ws_handler.py -> utils/kubectl/agent_ws_handler.py
- internal_api_handler.py -> utils/internal/api_handler.py

This improves code organization by grouping related utilities together and
keeping only main entry points (main_chatbot.py, main_compute.py,
celery_config.py) at the server root level.

Updated all imports in main_chatbot.py and api_handler.py to reference new
locations. Removed redundant COPY directives from Dockerfile since these
files are now included via the utils/ directory copy.
